Loading both new and old atlas via flag
